\name{Sharks}
\docType{data}
\alias{Sharks}
\title{Species richness of marine sharks and environmental variables}
\description{Species richness of sharks and environmental variables in cells of 60 minutes x 60 minutes around the world.}
\usage{data(Sharks)}
\format{A matrix of shark species richness in cells of 60 minutes x 60 minutes and the preveiling mean values of sea surface temperature, nitrate, salinity, depth and mean area of occupancy (AOO index) in each cell around the world.}
\source{The range maps of the species were elaborated following the methodology described by Guisande et al. (2013) and these maps were inputted into ModestR. The Bio-ORACLE global dataset 90_N 90_S real values were used as environmental factors (Tyberghein et al., 2012). The variables used were sea surface temperature, nitrate, salinity and depth. In addition to the variables mentioned above, the AOO index was also included. The output file from ModestR with the species richness and the environmental variables is a CSV file which was converted to a RData file.}
\references{
García-Roselló, E., Guisande, C., González-Dacosta, J., Heine, J., Pelayo-Villamil, P., Manjarrés-Hernández, A., Vaamonde, A. & Granado-Lorencio, C. (2013) ModestR: a software tool for managing and analyzing species distribution map databases. \emph{Ecography}, 36: 1202-1207.

Guisande, C., Patti, B., Vaamonde, A., Manjarrés-Hernández, A., Pelayo-Villamil, P., García-Roselló, E.,  González-Dacosta, J., Heine, J. & Granado-Lorencio, C. (2013) Factors affecting species richness of marine elasmobranchs. \emph{Biodiversity and Conservation}, 22: 1703-1714.

Pelayo-Villamil, P., Guisande, C., González-Vilas. L., Carvajal-Quintero. J.D., Jiménez-Segura, L.F., García-Roselló, E., Heine, J., González-Dacosta, J., Manjarrés-Hernández, A., Vaamonde, A., Granado-Lorencio, C. (2012) ModestR: Una herramienta infromática para el estudio de los ecosistemas acuáticos de Colombia. \emph{Actualidades Biológicas}, 34: 225-239.

Tyberghein L, Verbruggen H, Pauly K, Troupin C, Mineur F, de Clerck O (2012) Bio-ORACLE: a global environmental dataset for marine species distribution modelling. \emph{Global Ecology and Biogeography}, 21: 272-281.
}
\keyword{Sharks}

 
